A lot of updates....
Got the GUI to work properly!
Change the model to an not include pca.
Debugging model, as originally its accuracy was very low (around 40%) -> problem fixed as due to picture quality and pca model.
Testing preformed: numbers 9 and 8 perform badly in deployed model, perhaps due to overtraining.
Next stage is to try with a more generalizable model.